requirements.txt
requests==2.31.0
python-dotenv==1.0.1
slack-sdk==3.27.1
openai==1.30.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9